+++ resources/views/tailwind/form-errors.blade.php
@@ -1,7 +1,5 @@
-@if($name)
-    @error($name)
-        <p {!! $attributes->merge(['class' => 'text-red-500 text-xs italic']) !!}>
-            {{ $message }}
-        </p>
-    @enderror
-@endif
+@error($name)
+    <p {!! $attributes->merge(['class' => 'text-red-500 text-xs italic']) !!}>
+        {{ $message }}
+    </p>
+@enderror