+++ resources/views/tailwind/form-submit.blade.php
@@ -1,5 +1,8 @@
 <div class="mt-6 flex items-center justify-between">
-    <button class="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded focus:outline-none focus:shadow-outline" type="submit">
-        {{ $label }}
+    <button {!! $attributes->merge([
+        'class' => 'bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded focus:outline-none focus:shadow-outline',
+        'type' => 'submit'
+    ]) !!}>
+        {!! trim($slot) ?: __('Submit') !!}
     </button>
 </div>