+++ tests/Feature/TranslationTest.php
@@ -8,11 +8,21 @@ use ProtoneMedia\LaravelFormComponents\Tests\TestCase;
 class TranslationTest extends TestCase
 {
     /** @test */
-    public function it_can_bind_a_target_to_the_form()
+    public function it_can_have_no_target_bound_to_the_form()
     {
         $this->registerTestRoute('translation');
 
         $this->visit('/translation')
+            ->seeElement('input[name="input[nl]"][value=""]')
+            ->seeElement('input[name="input[en]"][value=""]');
+    }
+
+    /** @test */
+    public function it_can_bind_a_target_to_the_form()
+    {
+        $this->registerTestRoute('translation-with-bind');
+
+        $this->visit('/translation-with-bind')
             ->seeElement('input[name="input[nl]"][value="hallo"]')
             ->seeElement('input[name="input[en]"][value="hello"]');
     }
@@ -20,13 +30,13 @@ class TranslationTest extends TestCase
     /** @test */
     public function it_shows_the_validation_errors_and_old_values_correctly()
     {
-        $this->registerTestRoute('translation', function (Request $request) {
+        $this->registerTestRoute('translation-with-bind', function (Request $request) {
             $request->validate([
                 'input.*' => 'min:5',
             ]);
         });
 
-        $this->visit('/translation')
+        $this->visit('/translation-with-bind')
             ->type('hoi', 'input[nl]')
             ->type('hey', 'input[en]')
             ->press('Submit')