+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to `laravel-form-components` will be documented in this file
 
+## 1.0.1 - 2020-07-18
+
+- added `hasErrorAndShow` method to components
+- use `$slot` and added `$attributes` to `form-button`
+- added `$attibutes` to `form-group` and `form-label`
+- fixed translation bug
+
 ## 1.0.0 - 2020-07-17
 
 - initial release